requirements.txt
pydub
speechrecognition
python-dotenv
openai-whisper
PyQt5
opencv-python